requirements.txt
face_recognition 
opencv-python
RPi.GPIO
